+++ levee/_/bundle.lua
@@ -1,6 +1,9 @@
-return function(path)
+local M = {}
+
+
+local function _assets(path)
 	local t = {}
-	table.insert(t, 'return {')
+	table.insert(t, '{')
 	local cmd = "find -L '"..path.."' -type f"
 	for file in io.popen(cmd):lines() do
 		table.insert(t, ('\n\t["%s"] = "'):format(file:sub(#path+1)))
@@ -16,5 +19,25 @@ return function(path)
 		f:close()
 	end
 	table.insert(t, " }")
-	return loadstring(table.concat(t))
+	return table.concat(t)
 end
+
+
+function M.assets(path)
+	return loadstring("return ".._assets(path))
+end
+
+
+function M.templates(path)
+	return loadstring(([[
+		local _ = require("levee")._
+		local assets = %s
+		for k, v in pairs(assets) do
+			assets[k] = _.template(v)
+		end
+		return assets
+	]]):format(_assets(path)))
+end
+
+
+return M