+++ tests/_/test_bundle.lua
@@ -5,6 +5,6 @@ return {
 	test_core = function()
 		local filename = debug.getinfo(1, 'S').source:sub(2)
 		local path = _.path.dirname(filename)
-		assert(_.bundle(path)()["/test_bundle.lua"])
+		assert(_.bundle.assets(path)()["/test_bundle.lua"])
 	end,
 }